pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.121.2",
    "passlib[bcrypt]>=1.7.4",
    "pwdlib>=0.3.0",
    "pydantic>=2.12.4",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
